Studied nil and pass by reference to try to solve some obstacles.
I'm not even close to finish it.
Studied nil because I had to be sure that null and nil was the same to adapt this code to dart. (nil and null are different things, I'll see what could I do then if I face a failure because of that).
Also studied pass by reference to see be sure that those dict objects I was passing were new ones or I was changing the object I was passing.
Now I'm considering an option that I wasn't because I was thinking in golang. I'll check how to declare methods inside the class (I forgot I wasn't programming in golang) to see how it works.